Composition Over Inheritance 1
Created a forwarding class which implements the Java 7 List interface
and a List class called TotalAddedList which extends the forwarding
class. TotalAddedList counts how many total items have been added to the
list, regardless if they have since been deleted.

Ran out of time to test tonight but wanted to commit and push this as a
means of ensuring that I get to it tomorrow.